Authentication middleware

+++ backend/src/controllers/cartController.ts
@@ -1,23 +1,35 @@
 import { PrismaClient, ProductsOnCarts } from "@prisma/client";
 import { Request, Response } from "express";
+import { AuthRequest } from "../constants/helpers";
 
 const prisma = new PrismaClient();
 
-const getCart = async (req: Request, res: Response) => {
+const getCart = async (req: AuthRequest, res: Response) => {
   const { id } = req.params;
+  const { userId } = req;
   const cart = await prisma.cart.findUnique({
     where: { id: +id },
-    include: { products: true },
+    include: { products: true, user: true },
   });
-  if (!cart) {
+  if (!cart || cart.user.id !== userId) {
     return res.status(404).json({ error: "Cart not found" });
   }
   res.json({ cart });
 };
 
-const addItem = async (req: Request, res: Response) => {
+const addItem = async (req: AuthRequest, res: Response) => {
   const { id } = req.params;
+  const { userId } = req;
   const { productId, quantity }: ProductsOnCarts = req.body;
+
+  const cart = await prisma.cart.findUnique({
+    where: { id: +id },
+    include: { user: true },
+  });
+  if (!cart || cart.user.id !== userId) {
+    return res.status(404).json({ error: "Cart not found" });
+  }
+
   if (!productId) {
     return res.status(400).json({ error: "productId field is required" });
   }
@@ -34,13 +46,24 @@ const addItem = async (req: Request, res: Response) => {
   res.json({ product });
 };
 
-const updateItem = async (req: Request, res: Response) => {
+const updateItem = async (req: AuthRequest, res: Response) => {
   const { quantity }: ProductsOnCarts = req.body;
+  const { cartId, productId } = req.params;
+  const { userId } = req;
+
+  const cart = await prisma.cart.findUnique({
+    where: { id: +cartId },
+    include: { user: true },
+  });
+  if (!cart || cart.user.id !== userId) {
+    return res.status(404).json({ error: "Cart not found" });
+  }
+
   if (quantity === 0) {
     await deleteItem(req, res);
     return;
   }
-  const { cartId, productId } = req.params;
+
   const product = await prisma.productsOnCarts.update({
     where: { cartId_productId: { cartId: +cartId, productId: +productId } },
     data: {
@@ -50,16 +73,35 @@ const updateItem = async (req: Request, res: Response) => {
   res.json({ product });
 };
 
-const deleteItem = async (req: Request, res: Response) => {
+const deleteItem = async (req: AuthRequest, res: Response) => {
   const { cartId, productId } = req.params;
+  const { userId } = req;
+
+  const cart = await prisma.cart.findUnique({
+    where: { id: +cartId },
+    include: { user: true },
+  });
+  if (!cart || cart.user.id !== userId) {
+    return res.status(404).json({ error: "Cart not found" });
+  }
+
   const product = await prisma.productsOnCarts.delete({
     where: { cartId_productId: { cartId: +cartId, productId: +productId } },
   });
   return res.json({ product });
 };
 
-const clearCart = async (req: Request, res: Response) => {
+const clearCart = async (req: AuthRequest, res: Response) => {
   const { id } = req.params;
+  const { userId } = req;
+
+  const cart = await prisma.cart.findUnique({
+    where: { id: +id },
+    include: { user: true },
+  });
+  if (!cart || cart.user.id !== userId) {
+    return res.status(404).json({ error: "Cart not found" });
+  }
   await prisma.productsOnCarts.deleteMany({ where: { cartId: +id } });
   res.sendStatus(200);
 };
